requirements.txt
--index-url https://download.pytorch.org/whl/cu117
torch
torchvision
torchaudio
fastfeedforward==0.2.1
tqdm==4.66.1
hydra-core==1.3.0
hydra-submitit-launcher
pandas==2.2.0
torchinfo==1.8.0
soundfile==0.12.1
polars==0.20.7
einops==0.7.0
mlflow==2.14.0
dagshub==0.3.28
pymoo==0.6.6.1
